+++ app/build.gradle
@@ -4,7 +4,7 @@ plugins {
 
 android {
     namespace 'com.duksung.android_software'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.duksung.android_software"
@@ -44,4 +44,5 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     implementation 'org.jsoup:jsoup:1.13.1'
+    implementation 'com.android.volley:volley:1.2.1'
 }
