--- app/src/main/java/com/duksung/android_software/WebActivity.java
@@ -1,60 +1,0 @@
-package com.duksung.android_software;
-
-import android.os.Bundle;
-import android.view.View;
-import android.webkit.WebResourceRequest;
-import android.webkit.WebSettings;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-import android.widget.Button;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.duksung.android_software.R;
-
-public class WebActivity extends AppCompatActivity {
-    WebView webView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_web);
-
-        webView = (WebView) findViewById(R.id.webView);
-        WebSettings webSet = webView.getSettings();
-        webSet.setJavaScriptEnabled(true);
-        webView.loadUrl("http://203.252.213.36:8080/FinalProject/advisorForm.jsp");
-
-        Button btn = (Button) findViewById(R.id.button_web);
-        btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                finish();
-            }
-        });
-        /*
-        class MyWebViewClient extends WebViewClient {
-            @Override
-            public boolean shouldOverrideUrlLoading(WebView view, WebResourceRequest request) {
-                return super.shouldOverrideUrlLoading(view, request);
-            }
-        }
-        webView.setWebViewClient(new MyWebViewClient());
-        */
-        webView.setWebViewClient(new WebViewClient() {
-            @Override
-            public boolean shouldOverrideUrlLoading(WebView view, WebResourceRequest request) {
-                return super.shouldOverrideUrlLoading(view, request);
-            }
-        });
-    }
-
-    @Override
-    public void onBackPressed() {
-        if(webView.canGoBack()){
-            webView.goBack();
-        }else{
-            super.onBackPressed();
-        }
-    }
-}
