--- app/src/main/res/layout/activity_web.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <WebView
-        android:id="@+id/webView"
-        android:layout_width="409dp"
-        android:layout_height="729dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent" >
-
-        <Button
-            android:id="@+id/button_web"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_x="217dp"
-            android:layout_y="95dp"
-            android:text="Button" />
-    </WebView>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
