+++ app/src/main/res/menu/menu.xml
@@ -4,13 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     tools:context=".MainActivity">
     <item
-        android:id="@+id/nativeApp"
-        android:title="NATIVE"
+        android:id="@+id/reservation"
+        android:title="예매/OTT"
         app:showAsAction="always"/>
 
-    <item
-        android:id="@+id/WebApp"
-        android:title="WEB"
-        app:showAsAction="always"/>
 
 </menu>
